backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<6
psycopg[binary]>=3.1,<4
django-filter>=24.0
django-cors-headers>=4.3
dj-database-url>=2.1
gunicorn>=22.0
